+++ src/Service/MariaDBService.php
@@ -14,6 +14,7 @@ use Castor\Docker\Service\Builder\ComposeBuilder;
 
 use function Castor\Docker\docker_compose;
 use function Castor\Docker\expose_service_port;
+use function Castor\Docker\interactive_context;
 use function Castor\context;
 
 class MariaDBService implements DatabaseServiceInterface
@@ -74,7 +75,7 @@ class MariaDBService implements DatabaseServiceInterface
         yield [
             'task' => new AsTask('client', $this->getName(), 'Open a mariadb session on the database'),
             'function' => function (): void {
-                docker_compose(['exec', $this->getName(), 'mariadb', '-u', 'root', '-p' . $this->rootPassword, $this->database], c: context()->toInteractive());
+                docker_compose(['exec', $this->getName(), 'mariadb', '-u', 'root', '-p' . $this->rootPassword, $this->database], c: interactive_context());
             },
         ];